Fixed warnings. Refs #12

+++ src/netedit/GNENetHelper.cpp
@@ -1950,7 +1950,7 @@ GNENetHelper::AttributeCarriers::getNumberOfGenericDatas() const {
     int counter = 0;
     // iterate over all generic datas
     for (const auto &genericDataTag : myGenericDatas) {
-        counter += genericDataTag.second.size();
+        counter += (int)genericDataTag.second.size();
     }
     return counter;
 }
